Fix compilation on 64bit Kubuntu Dapper
svn path=/trunk/KDE/kdelibs/; revision=650862

+++ kfile/kfileplacesmodel.cpp
@@ -337,7 +337,7 @@ Qt::ItemFlags KFilePlacesModel::flags(const QModelIndex &index) const
 
 static QString _k_internalMimetype(const KFilePlacesModel * const self)
 {
-    return QString("application/x-kfileplacesmodel-")+QString::number((int)self);
+    return QString("application/x-kfileplacesmodel-")+QString::number( (long)self );
 }
 
 QStringList KFilePlacesModel::mimeTypes() const
